requirements.txt
aiogram==3.1.2
requests==2.31.0
pillow==10.0.0
pytesseract==0.3.10
openai==1.8.0
python-dotenv==1.0.0